Move model.util to model.string
