+++ src/main/java/net/sf/jabref/logic/formatter/bibtexfields/HtmlToLatexFormatter.java
@@ -8,7 +8,7 @@ import java.util.regex.Pattern;
 import net.sf.jabref.logic.formatter.Formatter;
 import net.sf.jabref.logic.l10n.Localization;
 import net.sf.jabref.logic.layout.LayoutFormatter;
-import net.sf.jabref.model.util.HTMLUnicodeConversionMaps;
+import net.sf.jabref.model.strings.HTMLUnicodeConversionMaps;
 
 import org.apache.commons.logging.Log;
 import org.apache.commons.logging.LogFactory;
