+++ src/main/java/net/sf/jabref/logic/openoffice/OOPreFormatter.java
@@ -3,7 +3,7 @@ package net.sf.jabref.logic.openoffice;
 import java.util.Map;
 
 import net.sf.jabref.logic.layout.LayoutFormatter;
-import net.sf.jabref.model.util.HTMLUnicodeConversionMaps;
+import net.sf.jabref.model.strings.HTMLUnicodeConversionMaps;
 import net.sf.jabref.model.strings.StringUtil;
 
 /**
