+++ src/main/java/net/sf/jabref/model/search/rules/RegexBasedSearchRule.java
@@ -6,7 +6,7 @@ import java.util.regex.Pattern;
 import java.util.regex.PatternSyntaxException;
 
 import net.sf.jabref.model.entry.BibEntry;
-import net.sf.jabref.model.util.LatexToUnicode;
+import net.sf.jabref.model.strings.LatexToUnicode;
 
 /**
  * Search rule for regex-based search.
